feat: cross_ref竞品反查 + MLVDS误标修复 + 速率标签真实值重构
- cross_ref(意图分类层方案A): 竞品型号确定性反查国产替代料, LLM只做意图分类,
  协议名(RS485/CAN/I2C)排除不当竞品型号, 去前缀黑名单改数据层判定
- MLVDS 5款(TPT9H/TPT9L)误标RS-485修复 + 协议互斥护栏(autofix.py)
- 速率标签梯子展开→单一真实值(方案甲): autofix/tag_config/query_parser/constraint-match
  搜索改数值≥比较(downgradable); 修中文数字"二十"→210解析bug
- ADC的MSPS/DAC的Settling Time误抽Mbps删除 + Mbps品类护栏(清97款脏标签)
- 测试: test_all 65/65 + test_constraint_layer 62/62 + build clean

+++ scripts/audit_data.py
@@ -92,7 +92,7 @@ def check_speed_unit(products):
         actual_mbps = kbps_val / 1000
         
         # Get max Mbps tag
-        mbps_tags = [int(t.replace('Mbps', '')) for t in feats.split() if t.endswith('Mbps')]
+        mbps_tags = [float(t.replace('Mbps', '')) for t in feats.split() if t.endswith('Mbps')]
         if not mbps_tags:
             continue
